Rename, reorg utils and update imports
